[Client] Avoid sending CellStates to server after initialization
Although b4e8560698fbe9ef30e1c5e004bf7f928e85c6a8 made players instantly able to see each other on minimalist servers that did not change their cells from the default, it created problems with the default CoreScripts where players need to be logged in before receipt of a CellStates packet from them is taken into account, with the result being that a player was recorded as having loaded their initial cells on the server's C++ side but not on the Lua side.

It may simply be best to expect servers to set player cells.